relocate dwarf section implementations to separate files as outer classes
