+++ substratevm/src/com.oracle.objectfile/src/com/oracle/objectfile/elf/dwarf/impl.java
@@ -1,0 +1,3 @@
+package com.oracle.objectfile.elf.dwarf;
+public class impl {
+}
